+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## [2.0.2]
+## [2.0.3] -2026/01/17
+- Sliders with moving average up to 11 days added for plot recovery and sleep charts in home tab
+-Dark mode always ON
+-Standardize y-axis scaling for consistent interpretation
+
+
+## [2.0.2] - 2026/01/16
 - Data freshness badge (workouts / sleep / recovery)
 - Home tab data status indicator (slightly delayed / fresh)